Adding Mirror Desktop Utility
A problem I noticed while using screen share to be able to control my ssytem is that if there are multiple monitors (2+) it is really hard to manage. To make  things simpler, you can mirror displays. Rather than force me to go through systme preferences each time, this adds a tool to do it for me